requirements.txt
annotated-types==0.7.0
anyio==4.12.0
bcrypt==5.0.0
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
defusedxml==0.7.1
Deprecated==1.3.1
Flask==3.1.2
Flask-Limiter==4.0.0
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
fonttools==4.61.0
fpdf==1.7.2
fpdf2==2.8.5
google-auth==2.43.0
google-genai==1.52.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.6.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ordered-set==4.1.0
packaging==25.0
pathlib==1.0.1
pillow==12.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyPDF2==3.0.1
python-dotenv==1.2.1
requests==2.32.5
rich==14.2.0
rsa==4.9.1
sniffio==1.3.1
SQLAlchemy==2.0.44
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.4
wrapt==2.0.1
